refactor: Standardize package name for view extensions
- Rename package `com.flpbrrs.mimigym.ui.extensions` to `com.flpbrrs.mimigym.ui.extension`.
- Update import path in `NewTrainingFragment.kt` to reflect the package name change for `hideKeyboard` extension.
